tests/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.20.0
requests>=2.28.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
fastapi>=0.68.0
websockets>=10.0
httpx>=0.24.0 